+++ bootstrapper/forms/formsNg2Bootstrapper.ts
@@ -6,7 +6,7 @@ import { services } from 'typescript-angular-utilities';
 import __timezone = services.timezone;
 
 import { FormComponent } from '../../source/components/form/form';
-import { ButtonSubmitComponent } from '../../source/components/buttonSubmit/buttonSubmit';
+import { ButtonSubmitComponent } from '../../source/components/buttons/buttonSubmit/buttonSubmit';
 import { CheckboxComponent } from '../../source/components/checkbox/checkbox';
 import { DateTimeComponent } from '../../source/components/dateTime/dateTime';
 import { TextboxComponent } from '../../source/components/textbox/textbox';
