+++ source/components/buttons/button/button.ng1.ts
