+++ source/components/buttons/buttonAsync/buttonAsync.ng1.ts
@@ -1,7 +1,7 @@
 ﻿import * as angular from 'angular';
 import * as _ from 'lodash';
 
-import { IPromiseUtility, serviceName as promiseServiceName, moduleName as promiseModuleName} from '../../services/promise/promise.service';
+import { IPromiseUtility, serviceName as promiseServiceName, moduleName as promiseModuleName} from '../../../services/promise/promise.service';
 
 import { buildButton, ButtonController } from '../button/button.ng1';
