+++ source/components/dateTime/dateTime.ts
@@ -12,7 +12,7 @@ import __date = services.date;
 import __dateFormats = __date.defaultFormats;
 import __timezone = services.timezone;
 
-import { ButtonComponent } from '../button/button';
+import { ButtonComponent } from '../buttons/button/button';
 import { ValidatedInputComponent, validationInputs, baseOutputs } from '../input/validationInput';
 import { ComponentValidator } from '../../services/componentValidator/componentValidator.service';
 import { FormComponent } from '../form/form';
